// app/src/generated/KvStoreClient.ts
export class KvStoreClient {
  private mero: MeroJs;
  private contextId: string;
  private executorPublicKey: string;

  constructor(mero: MeroJs, contextId: string, executorPublicKey: string) {
    this.mero = mero;
    this.contextId = contextId;
    this.executorPublicKey = executorPublicKey;
  }

  /**
   * init
   */
  public async init(): Promise<void> {
    const response = await this.mero.rpc.execute({ contextId: this.contextId, method: 'init', argsJson: {}, executorPublicKey: this.executorPublicKey });
    return response as void;
  }

  /**
   * set
   */
  public async set(params: { key: string; value: string }): Promise<void> {
    const response = await this.mero.rpc.execute({ contextId: this.contextId, method: 'set', argsJson: params, executorPublicKey: this.executorPublicKey });
    return response as void;
  }

  /**
   * update_if_exists
   */
  public async updateIfExists(params: { key: string; value: string }): Promise<boolean> {
    const response = await this.mero.rpc.execute({ contextId: this.contextId, method: 'update_if_exists', argsJson: params, executorPublicKey: this.executorPublicKey });
    return response as boolean;
  }

  /**
   * get_or_insert
   */
  public async getOrInsert(params: { key: string; value: string }): Promise<string> {
    const response = await this.mero.rpc.execute({ contextId: this.contextId, method: 'get_or_insert', argsJson: params, executorPublicKey: this.executorPublicKey });
    return response as string;
  }

  /**
   * entries
   */
  public async entries(): Promise<Record<string, string>> {
    const response = await this.mero.rpc.execute({ contextId: this.contextId, method: 'entries', argsJson: {}, executorPublicKey: this.executorPublicKey });
    return response as Record<string, string>;
  }

  /**
   * len
   */
  public async len(): Promise<number> {
    const response = await this.mero.rpc.execute({ contextId: this.contextId, method: 'len', argsJson: {}, executorPublicKey: this.executorPublicKey });
    return response as number;
  }

  /**
   * get
   */
  public async get(params: { key: string }): Promise<string> {
    const response = await this.mero.rpc.execute({ contextId: this.contextId, method: 'get', argsJson: params, executorPublicKey: this.executorPublicKey });
    return response as string;
  }

  /**
   * get_unchecked
   */
  public async getUnchecked(params: { key: string }): Promise<string> {
    const response = await this.mero.rpc.execute({ contextId: this.contextId, method: 'get_unchecked', argsJson: params, executorPublicKey: this.executorPublicKey });
    return response as string;
  }

  /**
   * get_result
   */
  public async getResult(params: { key: string }): Promise<string> {
    const response = await this.mero.rpc.execute({ contextId: this.contextId, method: 'get_result', argsJson: params, executorPublicKey: this.executorPublicKey });
    return response as string;
  }

  /**
   * remove
   */
  public async remove(params: { key: string }): Promise<string> {
    const response = await this.mero.rpc.execute({ contextId: this.contextId, method: 'remove', argsJson: params, executorPublicKey: this.executorPublicKey });
    return response as string;
  }

  /**
   * clear
   */
  public async clear(): Promise<void> {
    const response = await this.mero.rpc.execute({ contextId: this.contextId, method: 'clear', argsJson: {}, executorPublicKey: this.executorPublicKey });
    return response as void;
  }

}